+++ changelog.pl
@@ -1,5 +1,9 @@
 #
+<<<<<<< HEAD
+# Authentic Theme 17.00 (https://github.com/qooob/authentic-theme)
+=======
 # Authentic Theme 16.01 (https://github.com/qooob/authentic-theme)
+>>>>>>> 26c36195a7bc42a58e36b30aed57642ba4b432c4
 # Copyright 2015 Ilia Rostovtsev <programming@rostovtsev.ru>
 # Licensed under MIT (https://github.com/qooob/authentic-theme/blob/master/LICENSE)
 #
@@ -14,6 +18,18 @@ our $__changelog
               </div>
               <div class="modal-body" style="font-weight: 300">
 
+<<<<<<< HEAD
+                <h4>Version 17.00 (October 11, 2015)</h4>
+
+                <ul>
+
+                  <li>Added <em>10</em> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/264#issuecomment-147094497" target="_blank">predefined color palettes</a> for left navigation menu and <em>3</em> for the right content page</li>
+                  <li>Added separate options for <em>Webmin</em>/<em>Usermin</em> to manage security notifications</li>
+                  <li>Added quick access to theme\'s <em>extensions</em> editor and <em>logo</em> control from search</li>
+                  <li>Improved code highlight in <em>Bind</em>, <em>Dovecot</em>, <em>SSH</em>, <em>Squid</em>, <em>ProFTPD</em>, <em>Spamassassin</em>, <em>Samba</em>, <em>fail2ban</em> and other modules</li>
+                  <li>Improved rendering on <code>Webmin->Webmin Users->Available Webmin modules</code> tab</li>
+                  <li>Fixed bugs <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/255" target="_blank">#255</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/256" target="_blank">#256</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/259" target="_blank">#259</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/261" target="_blank">#261</a>
+=======
                 <h4>Version 16.00 - 16.01 (September 27, 2015)</h4>
 
                 <ul>
@@ -46,6 +62,7 @@ our $__changelog
                   <li>Fixed bugs <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/234" target="_blank">#234</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/235" target="_blank">#235</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/236" target="_blank">#236</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/238" target="_blank">#238</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/239" target="_blank">#239</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/242" target="_blank">#242</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/243" target="_blank">#243</a> <a class="label label-default" href="https://github.com/qooob/authentic-theme/issues/244" target="_blank">#244</a></li>
                   <li>Fixed other bugs and made dozes of improvements</li>
 
+>>>>>>> 26c36195a7bc42a58e36b30aed57642ba4b432c4
 
                 </ul>
                 <!-- <h5 style="margin-top:10px; margin-bottom:5px; font-weight: 300">How do I make the theme load faster?</h5> <em>Webmin</em> has in-built option to compress static files. It\'s possible to enable this feature by going to <code>Webmin->Webmin Configuration->Web Server Options</code> and setting <em>Gzip compress static files</em> option to <em>Use pre-compressed file and compress dynamically</em>.-->