+++ lib/JSON.pm
@@ -1,5 +1,4 @@
 package JSON;
-
 use strict;
 use Carp ();
 use base qw(Exporter);